feat: Update session titles in Vietnamese and add cancelled status to constants

+++ lib/src/viewmodel/growth_stage/growth_stage_state.dart
@@ -14,7 +14,7 @@ class GrowthStageState with _$GrowthStageState {
   const factory GrowthStageState.getRecommendedWeightByCageIdInProgress() =
       _GetRecommendedWeightByCageIdInProgress;
   const factory GrowthStageState.getRecommendedWeightByCageIdSuccess(
-          double recommendedWeight, List<int> weightList) =
+          double recommendedWeight, List<double> weightList) =
       _GetRecommendedWeightByCageIdSuccess;
   const factory GrowthStageState.getRecommendedWeightByCageIdFailure(
       String error) = _GetRecommendedWeightByCageIdFailure;
